+++ src/components/__tests__/common/DModal.spec.js
