+++ src/components/common/Banner.vue
