+++ src/components/common/GoTop.vue
